+++ src/index.ts
@@ -12,7 +12,7 @@ export { RecursivePartial } from './utils/commons';
 export { CurveType } from './utils/curves';
 export { timeFormatter, niceTimeFormatter, niceTimeFormatByDay } from './utils/data/formatters';
 export { DataGenerator } from './utils/data_generators/data_generator';
-export { DataSeriesColorsValues } from './lib/series/series';
+export { DataSeriesColorsValues } from './types/xy_chart/utils/series';
 export {
   AnnotationDomainType,
   AnnotationDomainTypes,
