+++ src/state/utils.test.ts
@@ -1,6 +1,6 @@
 import { mergeYCustomDomainsByGroupId } from '../types/xy_chart/utils/axis_utils';
 import { IndexedGeometry } from '../types/xy_chart/rendering/rendering';
-import { DataSeriesColorsValues, findDataSeriesByColorValues, getSeriesColorMap } from '../lib/series/series';
+import { DataSeriesColorsValues, findDataSeriesByColorValues, getSeriesColorMap } from '../types/xy_chart/utils/series';
 import {
   AreaSeriesSpec,
   AxisSpec,
