+++ src/types/xy_chart/tooltip/tooltip.ts
@@ -1,6 +1,6 @@
 import { TooltipValue } from '../utils/interactions';
 import { IndexedGeometry } from '../rendering/rendering';
-import { getColorValuesAsString } from '../../../lib/series/series';
+import { getColorValuesAsString } from '../utils/series';
 import { AxisSpec, BasicSeriesSpec } from '../utils/specs';
 
 export function getSeriesTooltipValues(tooltipValues: TooltipValue[]): Map<string, any> {
