+++ src/types/xy_chart/utils/series.ts
@@ -1,11 +1,11 @@
-import { ColorConfig } from '../../utils/themes/theme';
-import { Accessor } from '../../utils/accessor';
-import { GroupId, SpecId } from '../../utils/ids';
-import { splitSpecsByGroupId, YBasicSeriesSpec } from '../../types/xy_chart/domains/y_domain';
-import { formatNonStackedDataSeriesValues } from '../../types/xy_chart/utils/nonstacked_series_utils';
-import { isEqualSeriesKey } from '../../types/xy_chart/utils/series_utils';
-import { BasicSeriesSpec, Datum, SeriesAccessors } from '../../types/xy_chart/utils/specs';
-import { formatStackedDataSeriesValues } from '../../types/xy_chart/utils/stacked_series_utils';
+import { ColorConfig } from '../../../utils/themes/theme';
+import { Accessor } from '../../../utils/accessor';
+import { GroupId, SpecId } from '../../../utils/ids';
+import { splitSpecsByGroupId, YBasicSeriesSpec } from '../domains/y_domain';
+import { formatNonStackedDataSeriesValues } from './nonstacked_series_utils';
+import { isEqualSeriesKey } from './series_utils';
+import { BasicSeriesSpec, Datum, SeriesAccessors } from './specs';
+import { formatStackedDataSeriesValues } from './stacked_series_utils';
 
 export interface RawDataSeriesDatum {
   /** the x value */
