+++ src/types/xy_chart/utils/series_utils.test.ts
@@ -1,6 +1,6 @@
 import { getSpecId } from '../../../utils/ids';
 import { GeometryId } from '../rendering/rendering';
-import { DataSeriesColorsValues } from '../../../lib/series/series';
+import { DataSeriesColorsValues } from './series';
 import { belongsToDataSeries, isEqualSeriesKey } from './series_utils';
 
 describe('Series utility functions', () => {
