+++ src/types/xy_chart/utils/specs.ts
@@ -11,7 +11,7 @@ import { Omit, RecursivePartial } from '../../../utils/commons';
 import { AnnotationId, AxisId, GroupId, SpecId } from '../../../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../../../utils/scales/scales';
 import { CurveType } from '../../../utils/curves';
-import { DataSeriesColorsValues } from '../../../lib/series/series';
+import { DataSeriesColorsValues } from './series';
 
 export type Datum = any;
 export type Rotation = 0 | 90 | -90 | 180;
